+++ packages/ipfs-topology/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.11.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.10.0...@ceramicnetwork/ipfs-topology@2.11.0-rc.0) (2023-01-11)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
+# [2.10.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.10.0-rc.0...@ceramicnetwork/ipfs-topology@2.10.0) (2023-01-05)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
+# [2.10.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.9.1...@ceramicnetwork/ipfs-topology@2.10.0-rc.0) (2022-12-29)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
+## [2.9.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.9.0...@ceramicnetwork/ipfs-topology@2.9.1) (2022-12-29)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
+# [2.9.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.9.0-rc.0...@ceramicnetwork/ipfs-topology@2.9.0) (2022-12-21)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
+# [2.9.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.8.0...@ceramicnetwork/ipfs-topology@2.9.0-rc.0) (2022-12-15)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
 # [2.8.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.8.0-rc.0...@ceramicnetwork/ipfs-topology@2.8.0) (2022-12-08)
 
 **Note:** Version bump only for package @ceramicnetwork/ipfs-topology